+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [3.0.1] - 2026-03-10
+
+### Added
+- Added enum typecast attributes for backed enums:
+  - `Sirix\Cycle\Extension\Typecast\Enum\IntegerEnumType`
+  - `Sirix\Cycle\Extension\Typecast\Enum\StringEnumType`
+
+
 ## [3.0.0] - 2026-03-09
 
 ### Added